app/infrastructure/requirements.txt
aiofiles==24.1.0
alembic==1.14.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
async-property==0.2.2
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dependency-injector==4.45.0
deprecation==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
loguru==0.7.3
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.1
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-mock==3.14.0
python-dotenv==1.0.1
python-keycloak==5.3.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
win32_setctime==1.2.0
